refactored/removed npm block generation to utils file.
