+++ lib/controllers/npm.js
@@ -2,7 +2,7 @@ const { Router } = require('express')
 const qs = require('qs')
 const { WebClient } = require("@slack/web-api")
 const client = new WebClient(process.env.SLACK_BOT_TOKEN)
-const { npmScraper } = require('../utils/npm-utils')
+const { npmScraper, generateNpmBlocks } = require('../utils/npm-utils')
 const verifyToken = require('../middleware/verify-token')
 
 module.exports = Router()
@@ -12,67 +12,24 @@ module.exports = Router()
         res.send()
         
         /*
-        Convert text string to url query format for fetch, then fetch results.
+        Convert text string to url query format for fetch, then fetch scraping results.
         */
         const rawTextInput = req.body.text
         const parsedQuery = qs.stringify({q: rawTextInput}, { format:'RFC1738' })
         
-
         const scrapedData = await npmScraper(parsedQuery)
  
+        /* 
+        Format all Blockit sections before sending message to channel.
 
-        const formattedBlocks = [
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": `+++++ \n Top 3 search results for *${req.body.text}* @ npmjs: \n +++++`
-                }
-            },
-            {
-                "type": "divider"
-            },
-        ]
+        TODO add ability to parse for # of results.
+        */
+        const parsedBlocks = await generateNpmBlocks(rawTextInput, scrapedData)
 
-        scrapedData.slice(0, 3).forEach(ele => {
-            formattedBlocks.push(
-                {
-                    "type": "section",
-                    "text": {
-                        "type": "mrkdwn",
-                        "text": `*${ele.pkg}*`
-                    }
-                }
-            )
-            formattedBlocks.push(
-                {
-                    "type": "section",
-                    "text": {
-                        "type": "mrkdwn",
-                        "text": `_${ele.description}_`
-                    },
-                    "accessory": {
-                        "type": "button",
-                        "text": {
-                            "type": "plain_text",
-                            "text": `Go To: ${ele.pkg}`,
-                            "emoji": true,
-                        },
-                        "url": `${ele.link}`,
-                        "style": "primary",
-                        "action_id": "button-action"
-                    }
-                },
-                {
-                    "type": "divider"
-                }
-            )
-        })
-
-        
-        client.chat.postMessage({
+        // Send formatted results to channel.
+        const postMessage = await client.chat.postMessage({
             channel: req.body.channel_id,
-            blocks: formattedBlocks    
+            blocks: parsedBlocks    
         })
         
     })
