+++ lib/utils/npm-utils.js
@@ -44,6 +44,61 @@ const npmScraper = async (query) => {
 
 }
 
+
+const generateNpmBlocks = async (query, data, numResults = 3) => {
+    const formattedBlocks = [
+        {
+            "type": "section",
+            "text": {
+                "type": "mrkdwn",
+                "text": `+++++ \n Top ${numResults} search results for *${query}* @ npmjs: \n +++++`
+            }
+        },
+        {
+            "type": "divider"
+        },
+    ]
+
+    data.slice(0, numResults).forEach(ele => {
+        formattedBlocks.push(
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": `*${ele.pkg}*`
+                }
+            }
+        )
+        formattedBlocks.push(
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": `_${ele.description}_`
+                },
+                "accessory": {
+                    "type": "button",
+                    "text": {
+                        "type": "plain_text",
+                        "text": `Go To: ${ele.pkg}`,
+                        "emoji": true,
+                    },
+                    "url": `${ele.link}`,
+                    "style": "primary",
+                    "action_id": "button-action"
+                }
+            },
+            {
+                "type": "divider"
+            }
+        )
+    })
+
+    return formattedBlocks
+}
+
+
 module.exports = {
     npmScraper,
+    generateNpmBlocks
 }
